+++ CHANGELOG.md
@@ -8,3 +8,19 @@ and write what you replaced it with
 -->
 
 ## Next release
+
+- Create class for Task and property
+  - included enum for states
+  - included equality operators
+- Create test for title parser
+- Create test for due date parser
+- Create test for properties parser
+- Create test for switches parser
+- Create test for task factory
+- Create functions
+  - parseTask
+  - non still implemented methods:
+    - parseTitle
+    - parseDueDate
+    - parseSwitches
+    - parseProperties